Change: Consolidate CLI parsing
Move all CLI parsing to dedicated modules. This allows for better
separation of concerns and easier testing.
