+++ pontos/nvd/cve_changes/__init__.py
@@ -3,11 +3,11 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import asyncio
-from argparse import ArgumentParser, Namespace
-
-import shtab
+from argparse import Namespace
 
 from pontos.nvd.cve_changes.api import CVEChangesApi
+
+from ._parser import parse_args
 
 __all__ = ("CVEChangesApi",)
 
@@ -21,26 +21,6 @@ async def query_changes(args: Namespace) -> None:
             start_index=args.start,
         ):
             print(cve)
-
-
-def parse_args() -> Namespace:
-    parser = ArgumentParser()
-    shtab.add_argument_to(parser)
-    parser.add_argument("--token", help="API key to use for querying.")
-    parser.add_argument("--cve-id", help="Get changes for a specific CVE")
-    parser.add_argument(
-        "--event-name", help="Get all CVE associated with a specific event name"
-    )
-    parser.add_argument(
-        "--number", "-n", metavar="N", help="Request only N CPEs", type=int
-    )
-    parser.add_argument(
-        "--start",
-        "-s",
-        help="Index of the first CPE to request.",
-        type=int,
-    )
-    return parser.parse_args()
 
 
 def main() -> None:
