+++ pontos/release/main.py
@@ -9,9 +9,10 @@ import sys
 from typing import NoReturn
 
 from pontos.git import GitError
-from pontos.release.parser import parse_args
 from pontos.terminal.null import NullTerminal
 from pontos.terminal.rich import RichTerminal
+
+from ._parser import parse_args
 
 
 def main(
